weather/requirements.txt
absl-py
grpcio-tools
kafka-python
pymongo
requests